refactor: 레거시 컨트롤러 제거
