+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/ModelAndView.java
@@ -1,5 +1,8 @@
 package webmvc.org.springframework.web.servlet;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -17,6 +20,15 @@ public class ModelAndView {
     public ModelAndView addObject(final String attributeName, final Object attributeValue) {
         model.put(attributeName, attributeValue);
         return this;
+    }
+
+    public void render(final HttpServletRequest request,
+                       final HttpServletResponse response) {
+        try {
+            view.render(model, request, response);
+        } catch (Exception e) {
+            throw new IllegalArgumentException("view를 렌더링하다 에러가 발생했습니다.");
+        }
     }
 
     public Object getObject(final String attributeName) {
