+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handler/annoationhandler/AnnotationHandlerAdapter.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.handler.tobe.annoationhandler;
+package webmvc.org.springframework.web.servlet.mvc.handler.annoationhandler;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
