--- mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handler/asis/ForwardController.java
@@ -1,20 +1,0 @@
-package webmvc.org.springframework.web.servlet.mvc.handler.asis;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.util.Objects;
-
-public class ForwardController implements Controller {
-
-    private final String path;
-
-    public ForwardController(final String path) {
-        this.path = Objects.requireNonNull(path);
-    }
-
-    @Override
-    public String execute(final HttpServletRequest request, final HttpServletResponse response) {
-        return path;
-    }
-}
